camera-intelligence/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
opencv-python>=4.8.0
ultralytics>=8.0.0
shapely>=2.0.0
httpx>=0.25.0
pyyaml>=6.0
numpy>=1.24.0
